Cleanup
- Use primitive instead of boxed double with unneeded cast
- Use standard method keyword order

git-svn-id: svn+ssh://svn.code.sf.net/p/omegat/svn/trunk@9903 b0d8beef-cb45-0410-a8e4-c0d495c3b779

+++ src/org/omegat/core/machinetranslators/MyMemoryMachineTranslate.java
@@ -75,7 +75,7 @@ public final class MyMemoryMachineTranslate extends AbstractMyMemoryTranslate {
         // Find the best Human translation if no MT translation is provided for
         // this text. If there is a MT translation, it will always take
         // precedence.
-        Double bestScore = (double) 0;
+        double bestScore = 0d;
         Map<String, Object> bestEntry = null;
         Map<String, Object> mtEntry = null;
